pyproject.toml
[project]
name = "wakemypc-desktop"
version = "1.0.0"
description = "Desktop companion for wakemypc.com -- manage Pico W transmitters from a GUI"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
authors = [
    { name = "wakemypc.com contributors" },
]
dependencies = [
    "pyserial>=3.5",
]

[project.scripts]
wakemypc-desktop = "wakemypc_desktop.main:main"

[project.urls]
Homepage = "https://wakemypc.com"
Repository = "https://github.com/wakemypc/wakemypc-desktop"
Issues = "https://github.com/wakemypc/wakemypc-desktop/issues"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
